Fix paths for CI config files
By listing the correct paths in the Makefile, we stop 'make' from
re-running the associated 'cp' and 'sed' commands each time. For
example, we were telling make that we would create a file called
`gitlab-ci/application.yml`. On each run, make saw the file did not
exist, and tried to re-create it. By saying the file is in
`gitlab-ci/config/application.yml` we no longer mislead make.